requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
selenium==4.7.2
python-dotenv==0.21.0
pymongo==4.3.3
black==21.12b0
schedule==1.1.0
requests===2.27.1
requests-file===1.5.1